refactor: extract peer business logic to a separate file
